+++ mysql_libs.py
@@ -15,6 +15,7 @@
         create_slv_array
         crt_cmd
         crt_srv_inst
+        disconnect
         fetch_db_dict
         fetch_logs
         fetch_slv
@@ -55,7 +56,7 @@ import version
 __version__ = version.__version__
 
 
-def analyze_tbl(server, dbn, tbl, **kwargs):
+def analyze_tbl(server, dbn, tbl):
 
     """Function:  analyze_tbl
 
@@ -75,7 +76,7 @@ def analyze_tbl(server, dbn, tbl, **kwargs):
     return server.col_sql(cmd)
 
 
-def change_master_to(mst, slv, **kwargs):
+def change_master_to(mst, slv):
 
     """Function:  change_master_to
 
@@ -110,7 +111,7 @@ def change_master_to(mst, slv, **kwargs):
     print("Changed Slave: {0} to new Master: {1}".format(slv.name, mst.name))
 
 
-def checksum(server, dbn, tbl, **kwargs):
+def checksum(server, dbn, tbl):
 
     """Function:  checksum
 
@@ -130,7 +131,7 @@ def checksum(server, dbn, tbl, **kwargs):
     return server.col_sql(cmd)
 
 
-def check_tbl(server, dbn, tbl, **kwargs):
+def check_tbl(server, dbn, tbl):
 
     """Function:  check_tbl
 
@@ -150,7 +151,7 @@ def check_tbl(server, dbn, tbl, **kwargs):
     return server.col_sql(cmd)
 
 
-def chg_slv_state(slaves, opt, **kwargs):
+def chg_slv_state(slaves, opt):
 
     """Function:  chg_slv_state
 
@@ -179,7 +180,7 @@ def chg_slv_state(slaves, opt, **kwargs):
         gen_libs.prt_msg("Error", "No option selected to stop/start rep.")
 
 
-def create_instance(cfg_file, dir_path, cls_name, **kwargs):
+def create_instance(cfg_file, dir_path, cls_name):
 
     """Function:  create_instance
 
@@ -205,7 +206,7 @@ def create_instance(cfg_file, dir_path, cls_name, **kwargs):
         rep_japd=cfg.__dict__.get("rep_japd", None))
 
 
-def create_slv_array(cfg_array, add_down=True, **kwargs):
+def create_slv_array(cfg_array, add_down=True):
 
     """Function:  create_slv_array
 
@@ -225,7 +226,9 @@ def create_slv_array(cfg_array, add_down=True, **kwargs):
         slv_inst = mysql_class.SlaveRep(
             slv["name"], slv["sid"], slv["user"], slv["japd"],
             os_type=getattr(machine, slv["serv_os"])(), host=slv["host"],
-            port=int(slv["port"]), defaults_file=slv["cfg_file"])
+            port=int(slv["port"]), defaults_file=slv["cfg_file"],
+            rep_user=slv.get("rep_user", None),
+            rep_japd=slv.get("rep_japd", None))
         slv_inst.connect()
 
         if add_down or slv_inst.conn:
@@ -234,7 +237,7 @@ def create_slv_array(cfg_array, add_down=True, **kwargs):
     return slaves
 
 
-def crt_cmd(server, prog_name, **kwargs):
+def crt_cmd(server, prog_name):
 
     """Function:  crt_cmd
 
@@ -261,7 +264,7 @@ def crt_cmd(server, prog_name, **kwargs):
             server.host, "-P", str(server.port)]
 
 
-def crt_srv_inst(cfg, path, **kwargs):
+def crt_srv_inst(cfg, path):
 
     """Function:  crt_srv_inst
 
@@ -282,7 +285,33 @@ def crt_srv_inst(cfg, path, **kwargs):
         defaults_file=svr.cfg_file)
 
 
-def fetch_db_dict(server, **kwargs):
+def disconnect(*args):
+
+    """Function:  disconnect
+
+    Description:  Disconnects a class database connection.  Will check to see
+        if an argument is an list and if so will loop on the array to
+        disconnect all connections and only disconnect those connections with
+        activate connections.  Will require a disconnect method within the
+        class.
+
+    Arguments:
+        (input) *arg -> One or more connection instances.
+
+    """
+
+    for server in args:
+        if isinstance(server, list):
+            for srv in server:
+                if srv.conn:
+                    srv.disconnect()
+
+        else:
+            if server.conn:
+                server.disconnect()
+
+
+def fetch_db_dict(server):
 
     """Function:  fetch_db_dict
 
@@ -297,7 +326,7 @@ def fetch_db_dict(server, **kwargs):
     return server.col_sql("show databases")
 
 
-def fetch_logs(server, **kwargs):
+def fetch_logs(server):
 
     """Function:  fetch_logs
 
@@ -312,7 +341,7 @@ def fetch_logs(server, **kwargs):
     return server.col_sql("show binary logs")
 
 
-def fetch_slv(slaves, slv_name, **kwargs):
+def fetch_slv(slaves, slv_name):
 
     """Function:  fetch_slv
 
@@ -344,7 +373,7 @@ def fetch_slv(slaves, slv_name, **kwargs):
     return slv, err_flag, err_msg
 
 
-def fetch_tbl_dict(server, dbn, tbl_type="BASE TABLE", **kwargs):
+def fetch_tbl_dict(server, dbn, tbl_type="BASE TABLE"):
 
     """Function:  fetch_tbl_dict
 
@@ -365,7 +394,7 @@ def fetch_tbl_dict(server, dbn, tbl_type="BASE TABLE", **kwargs):
     return server.col_sql(qry)
 
 
-def find_name(slaves, name, **kwargs):
+def find_name(slaves, name):
 
     """Function:  find_name
 
@@ -387,7 +416,7 @@ def find_name(slaves, name, **kwargs):
     return None
 
 
-def is_cfg_valid(servers, **kwargs):
+def is_cfg_valid(servers):
 
     """Function:  is_cfg_valid
 
@@ -424,7 +453,7 @@ def is_cfg_valid(servers, **kwargs):
     return status, status_msg
 
 
-def is_logs_synced(mst, slv, **kwargs):
+def is_logs_synced(mst, slv):
 
     """Function:  is_logs_synced
 
@@ -452,7 +481,7 @@ def is_logs_synced(mst, slv, **kwargs):
     return is_synced
 
 
-def is_rep_delay(mst, slv, opt, **kwargs):
+def is_rep_delay(mst, slv, opt):
 
     """Function:  is_rep_delay
 
@@ -478,7 +507,7 @@ def is_rep_delay(mst, slv, opt, **kwargs):
     return is_delay
 
 
-def _io_rep_chk(mst, slv, is_delayed=False, **kwargs):
+def _io_rep_chk(mst, slv, is_delayed=False):
 
     """Function:  _io_rep_chk
 
@@ -505,7 +534,7 @@ def _io_rep_chk(mst, slv, is_delayed=False, **kwargs):
     return is_delayed
 
 
-def _sql_rep_chk(mst, slv, is_delayed=False, **kwargs):
+def _sql_rep_chk(mst, slv, is_delayed=False):
 
     """Function:  _sql_rep_chk
 
@@ -532,7 +561,7 @@ def _sql_rep_chk(mst, slv, is_delayed=False, **kwargs):
     return is_delayed
 
 
-def optimize_tbl(server, dbn, tbl, **kwargs):
+def optimize_tbl(server, dbn, tbl):
 
     """Function:  optimize_tbl
 
@@ -552,7 +581,7 @@ def optimize_tbl(server, dbn, tbl, **kwargs):
     return server.col_sql(cmd)
 
 
-def purge_bin_logs(server, prg_type, cutoff, **kwargs):
+def purge_bin_logs(server, prg_type, cutoff):
 
     """Function:  purge_bin_logs
 
@@ -571,7 +600,7 @@ def purge_bin_logs(server, prg_type, cutoff, **kwargs):
     server.cmd_sql(cmd)
 
 
-def reset_master(server, **kwargs):
+def reset_master(server):
 
     """Function:  reset_master
 
@@ -585,7 +614,7 @@ def reset_master(server, **kwargs):
     server.cmd_sql("reset master")
 
 
-def reset_slave(server, **kwargs):
+def reset_slave(server):
 
     """Function:  reset_slave
 
@@ -599,7 +628,7 @@ def reset_slave(server, **kwargs):
     server.cmd_sql("reset slave all")
 
 
-def select_wait_until(server, gtid_pos, timeout=0, **kwargs):
+def select_wait_until(server, gtid_pos, timeout=0):
 
     """Function:  select_wait_until
 
@@ -674,7 +703,7 @@ def start_slave_until(slv, log_file=None, log_pos=None, **kwargs):
     return err_flag, err_msg
 
 
-def switch_to_master(mst, slv, timeout=0, **kwargs):
+def switch_to_master(mst, slv, timeout=0):
 
     """Function:  switch_to_master
 
@@ -706,7 +735,7 @@ def switch_to_master(mst, slv, timeout=0, **kwargs):
     return status_flag
 
 
-def sync_delay(mst, slv, opt, **kwargs):
+def sync_delay(mst, slv, opt):
 
     """Function:  sync_delay
 
@@ -743,7 +772,7 @@ def sync_delay(mst, slv, opt, **kwargs):
             slv.stop_slave()
 
 
-def _io_delay_chk(mst, slv, **kwargs):
+def _io_delay_chk(mst, slv):
 
     """Function:  _io_delay_chk
 
@@ -773,7 +802,7 @@ def _io_delay_chk(mst, slv, **kwargs):
             print("Error: %s" % (err_msg))
 
 
-def sync_rep_slv(mst, slv, **kwargs):
+def sync_rep_slv(mst, slv):
 
     """Function:  sync_rep_slv
 
@@ -841,7 +870,7 @@ def wait_until(slv, opt, log_file=None, log_pos=None, **kwargs):
         _sql_wait_chk(slv, gtid, log_file, log_pos)
 
 
-def _io_wait_chk(slv, gtid, log_file, log_pos, **kwargs):
+def _io_wait_chk(slv, gtid, log_file, log_pos):
 
     """Function:  _io_wait_chk
 
@@ -877,7 +906,7 @@ def _io_wait_chk(slv, gtid, log_file, log_pos, **kwargs):
         slv.upd_slv_status()
 
 
-def _sql_wait_chk(slv, gtid, log_file, log_pos, **kwargs):
+def _sql_wait_chk(slv, gtid, log_file, log_pos):
 
     """Function:  _sql_wait_chk