+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [5.0.4] - 2021-02-05
+### Changed
+- mysql_libs.create_slv_array:  Added replication user information to mysql_class.SlaveRep instance call.
+- mysql_libs:  Removed unnesscary \*\*kwargs from argument lists.
+- mysql_class.MasterRep.connect, mysql_class.SlaveRep.connect:  Added silent option connect method call.
+
+### Added
+- mysql_libs.disconnect:  Disconnects one or more class database connections.
+
+
 ## [5.0.3] - 2021-01-07
 ### Changed
 - mysql_libs.create_instance:  Added "rep_user" and "rep_japd" to be allowed to be passed to class instance if detected.